+++ src/test/resources/testData/completion/groovy/EntityFieldCompletionOnEntityWithExtends.groovy
@@ -1,0 +1,4 @@
+package org.apache.ofbiz
+
+org.apache.ofbiz.entity.GenericValue bossOfMe = from('EntityFieldCompletionOnEntityWithExtends')
+def foo = bossOfMe.<caret>